+++ server/game/gamesteps/actionwindow.js
@@ -16,13 +16,6 @@ class ActionWindow extends UiPrompt {
         }
         this.prevPlayerPassed = false;
         this.game.currentActionWindow = this;
-        
-        if(!this.currentPlayer.promptedActionWindows[this.windowName]) {
-            this.game.addMessage('{0} has chosen to pass', this.currentPlayer);
-            this.prevPlayerPassed = true;
-            this.nextPlayer();
-        }
-        
     }
     
     activeCondition(player) {
@@ -30,6 +23,10 @@ class ActionWindow extends UiPrompt {
     }
 
     continue() {
+        if(!this.currentPlayer.promptedActionWindows[this.windowName]) {
+            this.pass();
+        }
+
         let completed = super.continue();
 
         if(!completed) {
@@ -65,7 +62,7 @@ class ActionWindow extends UiPrompt {
 
         if(choice === 'manual') {
             this.game.promptForSelect(this.currentPlayer, {
-                source: 'Play Action',
+                source: 'Manual Action',
                 activePrompt: 'Which ability are you using?',
                 cardCondition: card => (card.controller === this.currentPlayer && !card.facedown),
                 onSelect: (player, card) => {
@@ -77,35 +74,29 @@ class ActionWindow extends UiPrompt {
             return true;
         }
         
-        this.game.addMessage('{0} has chosen to pass', this.currentPlayer);
-        
-        if(this.prevPlayerPassed) {
-            this.complete();
+        if(choice === 'pass') {
+            this.pass();
             return true;
+        }
+
+    }
+
+    pass() {
+        this.game.addMessage('{0} passes', this.currentPlayer);
+        
+        if(this.prevPlayerPassed || !this.currentPlayer.otherPlayer) {
+            this.complete();
         }
 
         this.prevPlayerPassed = true;
         this.nextPlayer();
-
-        return true;
-    }
+   }
     
     nextPlayer() {
         let otherPlayer = this.game.getOtherPlayer(this.currentPlayer);
         
         if(otherPlayer) {
-            if(!otherPlayer.promptedActionWindows[this.windowName]) {
-                this.game.addMessage('{0} has chosen to pass', this.currentPlayer);
-                if(this.prevPlayerPassed) {
-                    this.complete();
-                } else {
-                    this.prevPlayerPassed = true;
-                }
-            } else {
-                this.currentPlayer = otherPlayer;
-            }
-        } else if(this.prevPlayerPassed) {
-            this.complete();
+            this.currentPlayer = otherPlayer;
         }
     }
