+++ server/game/gamesteps/conflict/conflictactionwindow.js
@@ -1,4 +1,4 @@
-const ActionWindow = require('./actionwindow.js');
+const ActionWindow = require('../actionwindow.js');
 
 const capitalize = {
     military: 'Military',
@@ -39,19 +39,14 @@ class ConflictActionWindow extends ActionWindow {
     }
     
     activePrompt() {
-        let buttons = [
-            { text: 'Pass', arg: 'pass' }
-        ];
-        if(this.game.manualMode) {
-            buttons.unshift({ text: 'Manual Action', arg: 'manual'});
-        }
-        
+        let props = super.activePrompt();
+
         this.conflict.calculateSkill();
         let conflictText = capitalize[this.conflict.conflictType] + ' ' + capitalize[this.conflict.conflictRing] + ' conflict';
         let skillText = 'Attacker: ' + this.conflict.attackerSkill + ' Defender: ' + this.conflict.defenderSkill;
         return {
             menuTitle: [conflictText, skillText].join('\n'),
-            buttons: buttons,
+            buttons: props.buttons,
             promptTitle: this.title
         };
     }
