+++ server/game/gamesteps/dynasty/dynastyactionwindow.js
@@ -4,10 +4,6 @@ const ActionWindow = require('../actionwindow.js');
 class DynastyActionWindow extends ActionWindow {
     constructor(game) {
         super(game, 'Play cards from provinces', 'dynasty');
-        if(!this.currentPlayer.promptedActionWindows[this.windowName]) {
-            this.pass(this.currentPlayer);
-            this.nextPlayer();
-        }
     }
 
     activePrompt() {
@@ -24,55 +20,26 @@ class DynastyActionWindow extends ActionWindow {
         };
     }
 
-    menuCommand(player, choice) {
-        player.canInitiateAction = false;
-
-        if(choice === 'manual') {
-            this.game.promptForSelect(this.currentPlayer, {
-                source: 'Manual Action',
-                activePrompt: 'Which ability are you using?',
-                cardCondition: card => (card.controller === this.currentPlayer && !card.facedown),
-                onSelect: (player, card) => {
-                    this.game.addMessage('{0} uses {1}\'s ability', player, card);
-                    this.markActionAsTaken();
-                    return true;
-                }
-            });
-            return true;
-        }
-        
-        if(choice === 'pass') {
-            this.pass(player);
-            this.nextPlayer();
-            return true;
-        }
-
-    }
-    
-    pass(player) {
-        if(_.all(this.game.getPlayers(), p => {
-            return p.passedDynasty === false;
-        })) {
-            this.game.addFate(this.currentPlayer, 1);
-            this.game.addMessage('{0} is the first to pass, and gains 1 fate.', player);
+    pass() {
+        this.currentPlayer.passDynasty();
+        if(!this.currentPlayer.opponent || !this.currentPlayer.opponent.passedDynasty) {
+            this.game.addMessage('{0} is the first to pass, and gains 1 fate.', this.currentPlayer);
+            this.game.raiseEvent('onFirstPassDuringDynasty', { player: this.currentPlayer }, () => this.game.addFate(this.currentPlayer, 1));
         } else {
-            this.game.addMessage('{0} has chosen to pass.', player);
+            this.game.addMessage('{0} passes.', this.currentPlayer);
         }
-        player.passDynasty();       
+
+        if(!this.currentPlayer.opponent || this.currentPlayer.opponent.passedDynasty) {
+            this.complete();
+        } else {
+            this.nextPlayer();
+        }
     }
     
     nextPlayer() {
         let otherPlayer = this.game.getOtherPlayer(this.currentPlayer);
         if(otherPlayer && !otherPlayer.passedDynasty) {
-            if(!otherPlayer.promptedActionWindows[this.windowName]) {
-                this.pass(otherPlayer);
-            } else {
-                this.currentPlayer = otherPlayer;
-            }
-        }
-        
-        if(this.currentPlayer.passedDynasty) {
-            this.complete();
+            this.currentPlayer = otherPlayer;
         }
     }
 }
